agent/requirements.txt
pika
psycopg2-binary
python-dotenv